refactor: Board 클래스 정적 팩토리 메소드 생성
