+++ src/main/java/chess/domain/Rank.java
@@ -1,6 +1,13 @@
 package chess.domain;
 
-import java.time.temporal.Temporal;
+import chess.domain.piece.Bishop;
+import chess.domain.piece.Blank;
+import chess.domain.piece.King;
+import chess.domain.piece.Knight;
+import chess.domain.piece.Pawn;
+import chess.domain.piece.Piece;
+import chess.domain.piece.Queen;
+import chess.domain.piece.Rook;
 import java.util.EnumMap;
 import java.util.Map;
